# Add `BracketedSort` a new, faster algorithm for `partialsort` and friends