--- a/base/sort.jl
+++ b/base/sort.jl
@@ -2152,6 +2152,7 @@ function uint_unmap!(v::AbstractVector, u::AbstractVector{U}, lo::Integer, hi::I
 end
 
 
+
 ### Unused constructs for backward compatibility ###
 
 ## Old algorithms ##