See the docstring of `BracketedSort` for an explanation of this algorithm.

The diff is kinda big but only 40% of it is code, the rest are comments, docstrings, and tests.

I produced it without reference to any sources that describe such an algorithm but I do not claim original authorship—somebody has certainly come up with this before.

Here's an earlier version: https://github.com/JuliaStats/Statistics.jl/issues/154

<details>
<summary>Robust benchmarks are available further down in the thread. Here's an appetizer</summary>

Some benchmarks vs master
```julia
julia> @btime partialsort!(x, 1:100) setup=(x=rand(Int, 100_000_000)) evals=1 samples=100 seconds=20;
  173.980 ms (3 allocations: 762.94 MiB)=>  65.699 ms (5 allocations: 4.49 MiB)

julia> @btime partialsort(x, 1:100) setup=(x=rand(Int, 100_000_000)) evals=1 samples=100 seconds=20;
  239.999 ms (6 allocations: 1.49 GiB)  => 132.764 ms (8 allocations: 767.43 MiB)

julia> @btime partialsortperm(x, 1:100) setup=(x=rand(Int, 100_000_000)) evals=1 samples=100 seconds=20;
  327.099 ms (8 allocations: 1.49 GiB)  => 321.387 ms (8 allocations: 1.49 GiB)

julia> @btime partialsort!(copy(x), 1:100) setup=(x=rand(10_000)) evals=1000 samples=1000 seconds=20;
  24.198 μs (6 allocations: 156.38 KiB) => 18.521 μs (6 allocations: 82.06 KiB)

julia> @btime partialsort(x, 1:100) setup=(x=rand(10_000)) evals=1000 samples=1000 seconds=20;
  18.803 μs (6 allocations: 156.38 KiB) => 18.105 μs (6 allocations: 82.06 KiB)

julia> @btime partialsortperm(x, 1:100) setup=(x=rand(10_000)) evals=1000 samples=1000 seconds=20;
  24.164 μs (8 allocations: 156.42 KiB) => 24.302 μs (8 allocations: 156.42 KiB)
```

A detailed view of a special case using `@benchmark` and including 1.9.3

```
julia> @benchmark partialsort!(x, 1:100) setup=(x=rand(Int, 100_000_000)) evals=1 samples=100 seconds=20

# PR
BenchmarkTools.Trial: 100 samples with 1 evaluation.
 Range (min … max):  68.893 ms … 111.625 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     69.499 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   70.898 ms ±   6.030 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▆▂                                                           
  ████▇▄▁▁▄▁▁▁▁▁▁▁▁▁▁▁▄▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄ ▄
  68.9 ms       Histogram: log(frequency) by time       103 ms <

 Memory estimate: 4.49 MiB, allocs estimate: 5.

# Master
BenchmarkTools.Trial: 49 samples with 1 evaluation.
 Range (min … max):  163.071 ms …    1.067 s  ┊ GC (min … max): 0.23% … 1.17%
 Time  (median):     216.896 ms               ┊ GC (median):    0.17%
 Time  (mean ± σ):   305.018 ms ± 204.787 ms  ┊ GC (mean ± σ):  0.20% ± 0.15%

  ▃▆█▆                                                           
  ████▇▇▇▄▁▆▁▄▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▄▄▁▁▁▁▄▄▁▁▁▁▄▆▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄ ▁
  163 ms           Histogram: frequency by time          1.07 s <

 Memory estimate: 762.94 MiB, allocs estimate: 3.

# 1.9.3
BenchmarkTools.Trial: 54 samples with 1 evaluation.
 Range (min … max):  161.129 ms … 894.473 ms  ┊ GC (min … max): 0.13% … 0.02%
 Time  (median):     213.916 ms               ┊ GC (median):    0.09%
 Time  (mean ± σ):   265.121 ms ± 142.888 ms  ┊ GC (mean ± σ):  0.08% ± 0.02%

     █▃▆                                                         
  ▇▄████▇▅▅▅▄▅▄▄▁▁▄▄▁▁▁▁▄▁▁▁▁▁▁▁▄▁▁▁▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▄ ▁
  161 ms           Histogram: frequency by time          742 ms <

 Memory estimate: 762.94 MiB, allocs estimate: 2.
```


</details>



I'm not touching the *perm varients because
- The cache locality makes me go "ewww!" and I don't want to touch them until I (or someone else) cleans that up
- When I tried I got wonky performance results including regressions that I don't want to investigate
- I want to pick good thresholds for `partialsortperm!` and I will only be able to do that once the overall approach is improved.

This makes `sortperm` unstable once again, with no option to make it stable. It was unstable in 1.9 with no option to make it stable, stable in 1.10 with no option to make it unstable, and has never been documented one way or the other.