@LilithHafner 2023-11-03
<details>
<summary>Here's a much more comprehensive benchmark</summary>

```julia
using Random, Printf
get_length() = round(Int, exp(17*rand()))
get_eltype() = rand([Float64, Float32, Int, UInt8, Int128, String])
get_function() = rand([partialsort!, partialsort])
get_element(x) = rand(x)
get_element(::Type{String}) = randstring()
get_k(n) = rand([
        1,
        n,
        rand(1:n),
        rand(1:1+n÷100),
        rand(n-n÷100:n),

        iseven(n) ? (n÷2:n÷2+1) : (n+1)÷2,

        UnitRange(sort((rand(1:n), rand(1:n)))...),
        1:rand(1:n),
        1:rand(1:1+n÷100),
        rand(n-n÷100:n):n,
        (1:rand(1:1+n÷100)) .+ rand(0:(n-n÷100-1)),
    ])

new!(v, k) = partialsort!(v, k)
new(v, k) = partialsort(v, k)
function old!(v, k)
    Base.Sort._sort!(v, Base.Sort.InitialOptimizations(Base.Sort.ScratchQuickSort(k)), Base.Order.Forward, (;))
    Base.Sort.maybeview(v, k)
end
old(v::AbstractVector, k::Union{Integer,OrdinalRange}; kws...) =
    old!(Base.copymutable(v), k; kws...)

function test(length::Int, eltype, func)
    func === partialsort && (length ÷= 2) # benchmarking is inefficient
    data = [get_element(eltype) for _ in 1:length]
    ks = [get_k(length) for _ in 1:5]
    sort!(ks, by = x->(x isa Number, x))
    unique!(ks)
    println("\n$func(rand($eltype, $length), k)")
    for k in ks
        (t_old, t_new) = func === partialsort! ? test!(similar(data), data, k) : test(data, k)
        a = 1e9t_old/length
        b = 1e9t_new/length
        c = log(b/a)*100
        @printf "k = %-18s %6.3f %6.3f  %+05.1f%%\n" k a b c
    end
end

estimate_runtime(data::Vector{T}, k) where T =
    round(Int, (10 + length(data) + length(k) * log(length(k))) * (T === String ? 3 : T == Int128 ? 2 : 1))
function test(data::Vector{T}, k::K) where {T,K}
    estimated_runtime = estimate_runtime(data, k)
    if estimated_runtime > 10^7
        t_old = @elapsed old_res = old(data, k)
        t_new = @elapsed new_res = new(data, k)
        old_res == new_res || error("Bad sort")
        t_old, t_new
    elseif estimated_runtime > 10^5
        truth = old(data, k)
        t_old, t_new = Inf, Inf
        for _ in 1:2*10^7 ÷ estimated_runtime
            t1 = @elapsed res1 = old(data, k)
            res1 == truth || error("Bad old sort")
            t_old = min(t_old, t1)

            t2 = @elapsed res2 = new(data, k)
            res2 == truth || error("Bad sort")
            t_new = min(t_new, t2)
        end
        t_old, t_new
    else
        old(data, k) == new(data, k) || error("Bad sort")
        evals = max(1, 10^4 ÷ estimated_runtime)
        samples = 10^7 ÷ evals ÷ estimated_runtime
        t_old = @belapsed old($data, $k) evals=evals samples=samples gctrial=false
        t_new = @belapsed new($data, $k) evals=evals samples=samples gctrial=false
        t_old, t_new
    end
end

function test!(d::Vector{T}, data::Vector{T}, k::K) where {T,K}
    estimated_runtime = estimate_runtime(data, k)
    if estimated_runtime > 10^7
        copyto!(d, data)
        t_old = @elapsed old_res = old!(d, k)
        copyto!(d, data)
        t_new = @elapsed new_res = new!(d, k)
        old_res == new_res || error("Bad sort")
    elseif estimated_runtime > 10^6
        truth = old(data, k)
        t_old, t_new = Inf, Inf
        for _ in 1:2*10^7 ÷ estimated_runtime
            copyto!(d, data)
            t1 = @elapsed res1 = old!(d, k)
            res1 == truth || error("Bad old sort")
            t_old = min(t_old, t1)

            copyto!(d, data)
            t2 = @elapsed res2 = new!(d, k)
            res2 == truth || error("Bad sort")
            t_new = min(t_new, t2)
        end
    else
        old!(copyto!(d, data), k) == new!(copyto!(d, data), k) || error("Bad sort")
        evals = max(1, 10^4 ÷ estimated_runtime)
        samples = 10^7 ÷ evals ÷ estimated_runtime
        if evals == 1
            t_old = @belapsed old($d, $k) setup=(copyto!($d, $data)) evals=evals samples=samples gctrial=false
            t_new = @belapsed new($d, $k) setup=(copyto!($d, $data)) evals=evals samples=samples gctrial=false
        else
            t_old = @belapsed old(copyto!($d, $data), $k) evals=evals samples=samples gctrial=false
            t_new = @belapsed new(copyto!($d, $data), $k) evals=evals samples=samples gctrial=false
        end
    end
    t_old, t_new
end

function test(n, seed=1729)
    Random.seed!(seed)
    samples = [(get_length(), get_eltype(), get_function()) for _ in 1:n]
    sort!(samples, by = x->(x[3] === partialsort, string(x[2]), x[1]))
    try
        for s in samples
            test(s...)
        end
    catch x
        x isa InterruptException || rethrow()
    end
    nothing
end

test(100)
```

</details>

<details>
<summary>And the results</summary>

The columns are `k` | old runtime | new runtime | change. Runtime is measured in nanoseconds per element. Changes are measured in logarithmic percent (a reported 1% change is a 1.01x or 0.99x, but a +100% is 2.71x and -100% is 0.37x) this makes positive and negative symmetric. `rand(String, 10)` is an abbreviation for `[randstring() for _ in 1:10]`.

```
partialsort!(rand(Float32, 3), k)
k = 1:1                11.595 17.104  +38.9%
k = 2:3                13.092 18.849  +36.5%
k = 2                  12.010 16.833  +33.8%
k = 3                  11.523 15.749  +31.2%

partialsort!(rand(Float32, 120), k)
k = 1:1                 2.207  2.632  +17.6%
k = 1:68                6.718  8.666  +25.5%
k = 120:120             3.148  4.162  +27.9%

partialsort!(rand(Float32, 356), k)
k = 1:1                 2.176  2.575  +16.8%
k = 178:179             2.822  3.377  +18.0%
k = 222:225             2.926  3.565  +19.8%
k = 1                   2.480  2.072  -18.0%
k = 356                 2.493  2.107  -16.8%

partialsort!(rand(Float32, 764), k)
k = 1:612               7.090  9.490  +29.2%
k = 69:736              7.526 10.034  +28.8%
k = 3                   2.227  1.636  -30.8%

partialsort!(rand(Float32, 1341), k)
k = 1:7                 3.773  2.250  -51.7%
k = 414:418             3.924  3.058  -24.9%
k = 683:1321            5.345  6.773  +23.7%
k = 1331                1.744  1.491  -15.7%
k = 1341                1.731  1.447  -17.9%

partialsort!(rand(Float32, 2417), k)
k = 2315:2339           1.954  1.787  -08.9%
k = 1                   2.129  1.228  -55.0%
k = 25                  2.133  1.302  -49.4%
k = 1209                2.383  3.599  +41.2%

partialsort!(rand(Float32, 1412894), k)
k = 706447:706448       2.456  1.606  -42.5%
k = 781870:784710       2.638  1.690  -44.5%
k = 849750:1250424      8.591 17.445  +70.8%
k = 265276              1.924  1.646  -15.6%
k = 1410938             1.768  1.091  -48.3%

partialsort!(rand(Float32, 3469856), k)
k = 1734928:1734929     3.008  1.761  -53.6%
k = 3441966:3469856     1.896  1.394  -30.7%
k = 2650041             1.752  1.737  -00.9%
k = 3461652             1.762  1.077  -49.2%
k = 3469697             1.774  1.047  -52.8%

partialsort!(rand(Float64, 3), k)
k = 1                  10.096 13.835  +31.5%
k = 2                  10.547 15.117  +36.0%
k = 3                   9.771 13.798  +34.5%

partialsort!(rand(Float64, 34), k)
k = 12:12               5.809  5.032  -14.4%
k = 1                   5.566  5.010  -10.5%
k = 5                   5.588  5.096  -09.2%
k = 32                  4.902  5.107  +04.1%

partialsort!(rand(Float64, 6994), k)
k = 1:843               3.390  3.289  -03.0%
k = 1:2826              6.077 10.396  +53.7%
k = 61                  2.693  1.364  -68.0%
k = 6707                1.638  1.531  -06.8%
k = 6994                1.567  1.418  -10.0%

partialsort!(rand(Float64, 1046517), k)
k = 161433:167648       2.730  2.151  -23.9%
k = 500996:870065      10.586 22.474  +75.3%
k = 6515                2.220  1.072  -72.8%
k = 351940              2.712  1.988  -31.1%
k = 1037448             2.267  1.292  -56.2%

partialsort!(rand(Float64, 1051336), k)
k = 1:4416              2.092  1.114  -63.0%
k = 455127:1037517     17.144 38.654  +81.3%
k = 1042624:1051336     5.153  1.371  -132.4%
k = 1                   2.063  0.947  -77.9%
k = 913974              4.896  1.893  -95.0%

partialsort!(rand(Float64, 3029236), k)
k = 1514618:1514619     3.359  1.977  -53.0%
k = 2999085:3029236     2.490  1.655  -40.8%
k = 160737              3.498  1.986  -56.6%
k = 3001131             2.621  1.981  -28.0%
k = 3029236             2.310  1.096  -74.6%

partialsort!(rand(Float64, 6212361), k)
k = 3649468:5431358    10.937 20.295  +61.8%
k = 6168786:6212361     2.770  1.464  -63.8%
k = 1                   2.592  1.079  -87.7%
k = 6212361             2.529  1.167  -77.4%

partialsort!(rand(Int128, 56), k)
k = 1:1                 5.784  8.571  +39.3%
k = 56:56               2.758  5.446  +68.0%
k = 16                  5.704  5.119  -10.8%
k = 50                  2.788  4.811  +54.6%
k = 56                  2.827  4.960  +56.2%

partialsort!(rand(Int128, 84), k)
k = 42:43               4.865  6.782  +33.2%
k = 62:62               3.650  5.840  +47.0%
k = 1                   3.875  4.305  +10.5%
k = 84                  3.182  2.630  -19.0%

partialsort!(rand(Int128, 271), k)
k = 1:2                 2.605  3.129  +18.3%
k = 108:110             3.283  3.898  +17.2%
k = 245:246             3.735  4.477  +18.1%
k = 271:271             4.016  4.766  +17.1%
k = 3                   2.614  2.686  +02.7%

partialsort!(rand(Int128, 4647), k)
k = 1                   4.035  2.035  -68.4%
k = 2270                4.430  2.385  -61.9%
k = 3279                4.178  2.457  -53.1%
k = 4644                1.740  1.766  +01.5%
k = 4647                1.757  1.784  +01.5%

partialsort!(rand(Int128, 6493), k)
k = 1:24                3.138  2.092  -40.5%
k = 1161:3576           7.900  7.887  -00.2%
k = 1722                2.971  2.394  -21.6%
k = 6493                1.906  1.668  -13.3%

partialsort!(rand(Int128, 11572), k)
k = 370:9931           18.839 18.244  -03.2%
k = 5786:5787           2.474  2.412  -02.5%
k = 11527:11572         2.675  1.718  -44.3%
k = 1602                2.744  2.247  -20.0%
k = 11528               2.679  1.617  -50.5%

partialsort!(rand(Int128, 24908), k)
k = 12454:12455         4.756  2.325  -71.6%
k = 21710               3.663  2.136  -53.9%
k = 24908               1.522  1.504  -01.2%

partialsort!(rand(Int128, 3802413), k)
k = 1:24959             3.391  1.685  -69.9%
k = 1:3621317          41.984 37.486  -11.3%
k = 1815638:1841498     4.307  1.962  -78.6%
k = 3789561:3802413     4.025  1.335  -110.3%
k = 3802413             3.904  1.226  -115.8%

partialsort!(rand(Int128, 5462941), k)
k = 1:21462             5.965  1.648  -128.7%
k = 1:5099526          39.881 38.254  -04.2%
k = 16598               6.804  1.565  -146.9%
k = 2731471             6.221  1.902  -118.5%
k = 5462941             6.516  1.262  -164.2%

partialsort!(rand(Int128, 22252660), k)
k = 1:90642            14.692  1.615  -220.8%
k = 1:3951357          11.860  8.545  -32.8%
k = 3032737:3090032     4.914  1.815  -99.6%
k = 1                   5.119  1.448  -126.3%
k = 22187610            4.722  1.216  -135.7%

partialsort!(rand(Int64, 4), k)
k = 1:1                 7.893  8.578  +08.3%
k = 4:4                 8.972  9.366  +04.3%
k = 4                   7.966  8.170  +02.5%

partialsort!(rand(Int64, 247), k)
k = 1                   2.406  2.530  +05.0%
k = 2                   1.558  2.135  +31.5%
k = 3                   2.215  2.131  -03.9%
k = 124                 3.476  2.357  -38.8%

partialsort!(rand(Int64, 885), k)
k = 96:381              5.446  5.948  +08.8%
k = 566:570             4.250  4.370  +02.8%
k = 878:885             1.605  2.076  +25.7%
k = 1                   1.331  1.558  +15.7%
k = 2                   1.391  1.806  +26.1%

partialsort!(rand(Int64, 9113), k)
k = 1:6373             10.813 10.644  -01.6%
k = 9081                1.280  1.559  +19.7%
k = 9103                1.262  1.084  -15.2%
k = 9113                1.257  1.079  -15.3%

partialsort!(rand(Int64, 66063), k)
k = 1:300               1.883  1.036  -59.8%
k = 36275:53130         5.594  5.700  +01.9%
k = 1                   1.706  0.887  -65.4%
k = 28981               2.393  1.278  -62.7%
k = 33032               1.647  1.279  -25.3%

partialsort!(rand(Int64, 80969), k)
k = 1:17084             5.396  4.470  -18.8%
k = 23302:70376        12.018 12.612  +04.8%
k = 33166:33845         2.900  1.404  -72.6%
k = 40                  1.894  0.913  -73.0%
k = 526                 1.894  0.918  -72.4%

partialsort!(rand(Int64, 170571), k)
k = 1:1038              1.144  1.022  -11.2%
k = 1:1629              1.177  1.025  -13.9%
k = 1:117683           15.013 14.777  -01.6%
k = 100458:145299       7.945  6.151  -25.6%
k = 85286               1.871  1.286  -37.5%

partialsort!(rand(Int64, 864519), k)
k = 2399:826475        23.491 23.279  -00.9%
k = 13365:242859        7.745  7.076  -09.0%
k = 203249:209420       2.948  1.701  -55.0%
k = 246445:252872       3.132  1.736  -59.0%
k = 432260              1.988  1.692  -16.1%

partialsort!(rand(String, 2), k)
k = 1:2                15.082 15.000  -00.5%
k = 1                  15.191 14.968  -01.5%
k = 2                  15.343 15.193  -01.0%

partialsort!(rand(String, 13), k)
k = 1:1                13.444 11.886  -12.3%
k = 1:12               13.023 11.193  -15.1%
k = 13:13              12.575 11.351  -10.2%
k = 1                  12.910 11.619  -10.5%
k = 13                 13.511 11.597  -15.3%

partialsort!(rand(String, 22), k)
k = 1:1                22.199 23.420  +05.4%
k = 11:12              22.083 23.769  +07.4%
k = 16:16              22.472 23.419  +04.1%
k = 22                  7.066 23.783  +121.4%

partialsort!(rand(String, 44), k)
k = 1:30               23.358 27.462  +16.2%
k = 34:34               8.259 11.860  +36.2%
k = 1                   8.321 12.730  +42.5%
k = 39                  8.259 13.506  +49.2%

partialsort!(rand(String, 138), k)
k = 1:1                15.508 16.140  +04.0%
k = 1:2                15.014 16.881  +11.7%
k = 1:111              28.005 29.438  +05.0%
k = 8:8                15.014 16.990  +12.4%
k = 1                  15.014  7.672  -67.1%

partialsort!(rand(String, 1159), k)
k = 1:8                 9.347  7.406  -23.3%
k = 1:10               10.443  7.370  -34.9%
k = 907:916            12.044 13.463  +11.1%
k = 10                 10.012  6.669  -40.6%
k = 890                12.116 11.558  -04.7%

partialsort!(rand(String, 1778), k)
k = 889:890            15.490 12.444  -21.9%
k = 1                   8.178  5.718  -35.8%
k = 4                   8.226  5.741  -36.0%
k = 1442                7.804 10.217  +26.9%
k = 1765                7.756  6.913  -11.5%

partialsort!(rand(String, 6996), k)
k = 1:9                 8.165  5.598  -37.7%
k = 480:500             9.636  8.541  -12.1%
k = 1143:6613          86.442 86.151  -00.3%
k = 2800:3871          30.410 24.085  -23.3%
k = 6996               10.447  5.664  -61.2%

partialsort!(rand(String, 9178), k)
k = 4589:4590          18.518 11.027  -51.8%
k = 7108:7329          18.082 11.772  -42.9%
k = 9112:9178          14.251  5.897  -88.2%
k = 1                   6.519  5.293  -20.8%
k = 9178               14.773  5.212  -104.2%

partialsort!(rand(String, 41621), k)
k = 19610:19752        19.027 10.953  -55.2%
k = 339                13.291  4.650  -105.0%
k = 20811              18.025 10.404  -55.0%
k = 29484              13.994  9.245  -41.5%
k = 41323              14.500  5.066  -105.2%

partialsort!(rand(String, 145977), k)
k = 135563:136325       7.751  9.293  +18.1%
k = 1088               11.616  4.787  -88.7%
k = 29000              16.266  8.101  -69.7%
k = 145913              5.996  4.815  -21.9%
k = 145977              5.875  4.828  -19.6%

partialsort!(rand(String, 6803515), k)
k = 1:3811305          181.665 184.072  +01.3%
k = 2206222:2240415    22.658 10.770  -74.4%
k = 6388849:6420065    22.926  9.106  -92.3%
k = 1581624            21.180  8.640  -89.7%
k = 1865233            19.512  8.863  -78.9%

partialsort!(rand(UInt8, 21), k)
k = 1:2                 5.456  4.547  -18.2%
k = 1                   5.675  4.523  -22.7%
k = 11                  5.539  4.504  -20.7%
k = 21                  5.607  4.523  -21.5%

partialsort!(rand(UInt8, 106), k)
k = 1:2                 1.535  2.904  +63.7%
k = 1:16                2.802  4.165  +39.6%
k = 1:77                5.235  6.700  +24.7%
k = 82:83               3.043  4.537  +39.9%
k = 34                  2.230  2.473  +10.3%

partialsort!(rand(UInt8, 71092), k)
k = 30602:31046         1.325  1.156  -13.6%
k = 35546:35547         1.281  1.185  -07.8%
k = 1                   0.797  0.554  -36.3%
k = 12198               0.753  1.015  +29.9%

partialsort!(rand(UInt8, 176199), k)
k = 175743:176199       1.066  0.677  -45.4%
k = 1                   0.850  0.589  -36.7%
k = 88100               1.389  1.141  -19.6%
k = 176199              1.028  0.723  -35.1%

partialsort!(rand(UInt8, 2649352), k)
k = 1:891915            6.867  6.232  -09.7%
k = 1324676:1324677     1.189  0.990  -18.3%
k = 1                   1.630  0.538  -110.8%
k = 687451              2.050  1.108  -61.5%
k = 2648225             0.916  0.653  -33.8%

partialsort!(rand(UInt8, 19380151), k)
k = 1:1974500           3.210  2.316  -32.7%
k = 1                   1.144  0.591  -66.1%
k = 9690076             1.424  1.101  -25.7%
k = 19315724            1.196  0.694  -54.4%

partialsort(rand(Float32, 42), k)
k = 11:12               4.654  8.016  +54.4%
k = 16:27               6.125 10.700  +55.8%
k = 11                  4.454  5.157  +14.6%
k = 42                  4.299  5.549  +25.5%

partialsort(rand(Float32, 49), k)
k = 1:10                5.924  8.946  +41.2%
k = 1:15                7.160 10.553  +38.8%
k = 1                   3.452  4.272  +21.3%
k = 49                  3.069  4.946  +47.7%

partialsort(rand(Float32, 254), k)
k = 3:34                2.726  3.003  +09.7%
k = 127:128             2.802  3.369  +18.4%
k = 252:254             2.350  2.687  +13.4%
k = 118                 2.908  3.064  +05.2%
k = 254                 2.345  2.696  +13.9%

partialsort(rand(Float32, 311), k)
k = 138:176             4.236  5.416  +24.6%
k = 169:170             3.570  4.486  +22.8%
k = 1                   2.213  2.070  -06.7%
k = 156                 3.375  2.273  -39.5%

partialsort(rand(Float32, 849), k)
k = 1:557               7.362  9.545  +26.0%
k = 448:453             3.467  2.802  -21.3%
k = 483:490             3.547  2.806  -23.4%
k = 1                   1.919  1.566  -20.3%
k = 848                 3.101  1.669  -62.0%

partialsort(rand(Float32, 1179), k)
k = 1:3                 2.748  1.807  -41.9%
k = 270:472             5.425  4.780  -12.7%
k = 837:921             3.935  3.540  -10.6%
k = 1                   2.774  1.529  -59.6%
k = 1179                1.427  1.577  +10.0%

partialsort(rand(Float32, 266404), k)
k = 133202:133203       1.984  1.654  -18.2%
k = 263867:266404       1.681  1.337  -22.9%
k = 266216:266404       1.494  1.024  -37.7%
k = 100179              1.954  1.650  -16.9%

partialsort(rand(Float32, 1769209), k)
k = 88264:771345       11.808 25.203  +75.8%
k = 999524:1000381      2.175  2.017  -07.6%
k = 1768923:1769209     2.472  1.319  -62.8%
k = 3977                2.974  1.242  -87.4%
k = 1769209             2.311  1.344  -54.2%

partialsort(rand(Float64, 4), k)
k = 2:2                 6.944  9.687  +33.3%
k = 2:3                 7.445 10.807  +37.3%
k = 4:4                 8.068  9.556  +16.9%
…